% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{bivariate_gaussian_laplace}
\alias{bivariate_gaussian_laplace}
\title{Bivariate Gaussian with Laplace margins}
\usage{
bivariate_gaussian_laplace(sample_n, rho)
}
\arguments{
\item{sample_n}{an integer, size of sample to be generated}

\item{rho}{float, correlation between the two variables}
}
\value{
matrix, sample with laplace margins and gaussian dependence structure
}
\description{
function which generates a bivariate Gaussian sample with laplace margins
}
