% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_dens.R
\name{box_prob}
\alias{box_prob}
\title{Estimates the probability of a point being in a box using HT (2004)}
\usage{
box_prob(point, fit, x_div, y_div, adjust = 0)
}
\arguments{
\item{point}{the centre point of the box}

\item{fit}{the fitted conditional extremes model}

\item{x_div}{the division points for x}

\item{y_div}{the division points for y}

\item{adjust}{inflation factor for residual noise kernel smoothing sd, defaults to 0. Making this positive can reduce 'streak' effects in extrapolation}
}
\description{
Uses a fitted conditional extremes (HT, 2004) model to estimate the probability of
lying within a box with given centre point
}
\keyword{internal}
