% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_and_pred.R
\name{case_one}
\alias{case_one}
\title{returns truth value of case one for constraints given by keef (2013)}
\usage{
case_one(theta, quantiles, v)
}
\arguments{
\item{theta}{H&T parameter}

\item{quantiles}{?? check Keef(2013) for this}

\item{v}{?? check keef(2013) for this}
}
\value{
boolean, true if case 1 satisfied
}
\description{
returns truth value of case one for constraints given by keef (2013)
}
\keyword{internal}
