% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_opt.R
\name{equality2}
\alias{equality2}
\title{second equality constraint for optimisation}
\usage{
equality2(beta, Yi, Ymnsi, alpha, v)
}
\arguments{
\item{beta}{conditional extremes parameter}

\item{Yi}{the excesses of the conditioning variable}

\item{Ymnsi}{the excesses of the conditioned variable}

\item{alpha}{conditional extremes parameter}

\item{v}{a large value chosen as part of Keef(2013) constraints}
}
\value{
the value of the equality constraint
}
\description{
second equality constraint for optimisation
}
\keyword{internal}
