% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cases.R
\name{gaussian_prob}
\alias{gaussian_prob}
\title{true probability of joint exceedance for a gaussian dependence structure}
\usage{
gaussian_prob(v, rho)
}
\arguments{
\item{v}{the joint exceedance value on laplace margins}

\item{rho}{the dependence parameter}
}
\value{
the true probability of joint exceedance
}
\description{
true probability of joint exceedance for a gaussian dependence structure
}
