% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_opt.R
\name{ht.fitAlt}
\alias{ht.fitAlt}
\title{alternative optimisation for H&T (2004) model over the constrained parameter space of Keef (2013)}
\usage{
ht.fitAlt(Yi, Ymnsi, qu)
}
\arguments{
\item{Yi}{the excesses of the conditioning variable}

\item{Ymnsi}{the excesses of the conditioned variable}

\item{qu}{the quantile of the conditioning variable}
}
\value{
list of the estimated parameters and residuals
}
\description{
optimisating in this way is faster, useful for cases where many repeated fits are needed (e.g., parametric bootstrapping)
}
