% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bivariate_dens.R
\name{ht.pred.cnstr}
\alias{ht.pred.cnstr}
\title{Predicting HT points within set xrange}
\usage{
ht.pred.cnstr(v, theta, Z, n_pred, dx, Y = T, points = F, plot = T, adjust = 0)
}
\arguments{
\item{v}{prediction threshold}

\item{theta}{ht model parameters}

\item{Z}{ht model residuals}

\item{n_pred}{number of predictions to make}

\item{dx}{width of restricted x range}

\item{Y}{true if doing Y|X, false if doing X|Y}

\item{points}{true to return prediction points in a matrix}

\item{plot}{true to plot predicted points}

\item{adjust}{inflation factor for residual noise sd, defaults to 0}
}
\value{
matrix of predicted points
}
\description{
Prediction of points using the ht(2004) model within a set range of x values
}
