% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/test_cases.R
\name{neglog_prob}
\alias{neglog_prob}
\title{true probability of joint exceedance for an inverted logistic dependence structure}
\usage{
neglog_prob(v, alpha)
}
\arguments{
\item{v}{the joint exceedance value on laplace margins}

\item{alpha}{the dependence parameter}
}
\value{
the true probability of joint exceedance
}
\description{
true probability of joint exceedance for an inverted logistic dependence structure
}
