% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{neglog_sample_lap}
\alias{neglog_sample_lap}
\title{MEVT sample with laplace margins generation}
\usage{
neglog_sample_lap(sample_n, alpha)
}
\arguments{
\item{sample_n}{an integer, size of sample to be generated}

\item{alpha}{floating point in \link{0,1}, MEVD inverted logistic dependence parameter}
}
\value{
matrix, sample with laplace margins and inverted logistic dep structure
}
\description{
function which generates a bivariate sample with MEVD inverted logistic dependence
structure and laplace margins
}
