% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{qspliced_v_fast}
\alias{qspliced_v_fast}
\title{Vectorised quantile function for GPD spliced with empirical}
\usage{
qspliced_v_fast(p, x, q, gpd_par)
}
\arguments{
\item{p}{vector of probabilities between \link{0,1} for which to find the corresponding quantile}

\item{x}{vector of sample used to fit the distribution}

\item{q}{float between \link{0,1}, exceedance quantile}

\item{gpd_par}{vector of gpd scale and shape parameters}
}
\value{
vector of quantiles, estimated pth quantile for data x with gpd tail
}
\description{
Vectorised quantile function for GPD spliced with empirical
}
\author{
Wanchen Yue
}
