% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alt_opt.R
\name{zM}
\alias{zM}
\title{maximum value of the residuals, as a function of beta, for a given alpha}
\usage{
zM(beta, Yi, Ymnsi, alpha)
}
\arguments{
\item{beta}{conditional extremes parameter}

\item{Yi}{the excesses of the conditioning variable}

\item{Ymnsi}{the excesses of the conditioned variable}

\item{alpha}{conditional extremes parameter}
}
\value{
the maximum value of the residuals as a function of beta, given alpha
}
\description{
This is needed to optimise and find boundary of Keef(2013) parameter space
}
\keyword{internal}
