% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prob_est.R
\name{prob_est}
\alias{prob_est}
\title{Probability estimation using the conditional extremes model}
\usage{
prob_est(data, v, q.cond, npred = 1e+05)
}
\arguments{
\item{data}{a matrix with two columns, the first being the x values and the second being the y values, on laplace margins}

\item{v}{the joint exceedance value on laplace margins}

\item{q.cond}{vector of conditioning quantiles}

\item{npred}{number of predictions to make in empirical probability estimation}
}
\value{
the estimated probability of joint exceedance
}
\description{
Probability estimation using the conditional extremes model
}
